backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
requests>=2.31.0
groq>=0.4.2
